remove iteneraries service and implement the method on chalenges.service
