+++ src/app/services/challenge.service.spec.ts
@@ -1,8 +1,12 @@
 import { ChallengeService } from "./challenge.service";
-import {TestScheduler} from "rxjs/internal/testing/TestScheduler";
-import {HttpTestingController} from "@angular/common/http/testing";
-import {delay, of} from "rxjs";
-import data from "../../assets/dummy/challenge.json"; 
+import { TestScheduler } from "rxjs/internal/testing/TestScheduler";
+import { HttpClientTestingModule, HttpTestingController } from "@angular/common/http/testing";
+import { delay, of } from "rxjs";
+import data from "../../assets/dummy/challenge.json";
+import { HttpClient } from "@angular/common/http";
+import { TestBed } from "@angular/core/testing";
+import { environment } from "src/environments/environment";
+import { Itinerary } from "../models/itinerary.interface";
 
 
 /* Observable Test, see https://docs.angular.lat/guide/testing-components-scenarios */
@@ -11,11 +15,20 @@ describe('ChallengeService', () => {
     let service: ChallengeService;
     let httpMock: HttpTestingController;
     let scheduler: TestScheduler;
+    let httpClient: HttpClient;
+    let httpClientMock: HttpTestingController;
     let httpClientSpy: any;
     let testScheduler: TestScheduler;
-    
+
 
     beforeEach(() => {
+        TestBed.configureTestingModule({ // set up the testing module with required dependencies.
+            imports: [HttpClientTestingModule]
+        });
+
+        // Inject the http service and test controller for each test
+        httpClient = TestBed.inject(HttpClient); //TestBed.inject is used to inject into the test suite
+        httpClientMock = TestBed.inject(HttpTestingController);
 
         //inject spy
         httpClientSpy = jasmine.createSpy('httpClient');
@@ -47,16 +60,63 @@ describe('ChallengeService', () => {
      */
 
     it('Should stream a challenge', () => {
-
-        testScheduler.run(({expectObservable}) => {
-            const idChallenge = "1adfadf21fasdf2-adf" 
+        testScheduler.run(({ expectObservable }) => {
+            const idChallenge = "1adfadf21fasdf2-adf"
             const expectedMarble = '---(a|)';
-            const expectedValues = {a: data};
+            const expectedValues = { a: data };
             const obs$ = service.getChallengeById(idChallenge).pipe(delay(3));
 
             expectObservable(obs$).toBe(expectedMarble, expectedValues);
         });
+    });
 
+    it('should be created itineraries.service', (done) => {
+        expect(service).toBeTruthy();
+        done();
+    });
+
+    it('should get itineraries succesfully', (done) => {
+        const mockData: Itinerary[] = [
+            {
+                id: '1',
+                name: 'mockName',
+                slug: 'mockSlug',
+            }
+        ];
+
+        httpClient.get<Itinerary[]>(environment.BACKEND_ITA_SSO_BASE_URL.concat(environment.BACKEND_SSO_ITINERARIES)).subscribe((res) => {
+            expect(res).toEqual(mockData);
+            done();
+        });
+
+        const req = httpClientMock.expectOne(environment.BACKEND_ITA_SSO_BASE_URL.concat(environment.BACKEND_SSO_ITINERARIES));
+        expect(req.request.method).toEqual('GET');
+        req.flush(mockData);
+
+    });
+
+    it('should handle error when getting itineraries', (done) => {
+
+        httpClient.get<Itinerary[]>(environment.BACKEND_ITA_SSO_BASE_URL.concat(environment.BACKEND_SSO_ITINERARIES)).subscribe({
+            next: () => {
+                done.fail('Expected error but received next');
+            },
+            error: (err) => {
+                expect(err).toBeTruthy();
+                done();
+            }
+        });
+
+        const req = httpClientMock.expectOne(environment.BACKEND_ITA_SSO_BASE_URL.concat(environment.BACKEND_SSO_ITINERARIES));
+        expect(req.request.method).toEqual('GET');
+
+        req.error(new ProgressEvent('error', {
+            lengthComputable: false,
+            loaded: 0,
+            total: 0,
+        }));
+
+        httpClientMock.verify();
     });
 
 });
