+++ src/app/services/challenge.service.ts
@@ -1,6 +1,8 @@
-import {Injectable} from "@angular/core";
-import {Observable} from "rxjs";
-import {HttpClient} from "@angular/common/http";
+import { Injectable } from "@angular/core";
+import { Observable } from "rxjs";
+import { HttpClient } from "@angular/common/http";
+import { Itinerary } from "../models/itinerary.interface";
+import { environment } from "src/environments/environment";
 //import {environment} from "../../environments/environment";
 
 
@@ -10,10 +12,10 @@ import {HttpClient} from "@angular/common/http";
 
 export class ChallengeService {
 
-    
-    constructor(private http: HttpClient) {}
-  
-    getChallengeById(id: string):Observable<Object>{
+
+    constructor(private http: HttpClient) { }
+
+    getChallengeById(id: string): Observable<Object> {
 
         //-----TO CHANGE----
 
@@ -31,5 +33,16 @@ export class ChallengeService {
                     'Content-Type': 'application/dummy'
                 }
             });
+    }
+
+
+    getItineraries(): Promise<Itinerary[]> {
+        return new Promise((resolve, reject) =>
+            this.http.get<Itinerary[]>(environment.BACKEND_ITA_SSO_BASE_URL.concat(environment.BACKEND_SSO_ITINERARIES))
+                .subscribe({
+                    next: (res) => resolve(res),
+                    error: (err) => reject([])
+                })
+        )
     }
 }
